Add initial migrations and create library endpoint

+++ internal/model/book.go
@@ -3,12 +3,12 @@ package model
 import "time"
 
 type Book struct {
-	ID string
-	Name string
-	URL string
-	TitleID string
-	LibraryID string
-	PageCount int64
-	CreatedDate time.Time
+	ID               string
+	Name             string
+	URL              string
+	TitleID          string
+	LibraryID        string
+	PageCount        int
+	CreatedDate      time.Time
 	LastModifiedDate time.Time
 }
